PythonScripts/requirements.txt
google-genai
Pillow
numpy
